test_push_notifications: Split AnalyticsBouncerTest to a separate file.
This commit moves `AnalyticsBouncerTest` to a new
`test_zilencer_analytics.py` file.

It helps in making it easier to work with `test_push_notifications.py`
which was 5000+ lines of code.
